Disable paths matching dialog, allow duplicate path tabs
If the path of the dialog window matches the path you put into the Edit text box to navigate to, it doesn't work. Not that you'd want to do that anyway. But now menu items for tabs matching the path will be disabled. Same for the clipboard item.

Also now all tabs will be shown in the menu for all windows even if they are the same path as another tab. Might add a setting to change this later.